+++ wookie/sbsdftransformers/__init__.py
@@ -1,1 +1,1 @@
-
+from wookie.sbsdftransformers.fuzzy import FuzzyWuzzySbsComparator
